deleting therapist favorites from web view

+++ src/pages/therapist/therapist.ts
@@ -39,11 +39,13 @@ export class TherapistPage {
   }
 
   studentTapped(student) {
-    this.navCtrl.push(TherapistStudentDetailsPage, {
-      student: student,
-      status: this.studentList.get(student).location,
-      id: this.id
-    });
+    if(!this.editDesktop) {
+      this.navCtrl.push(TherapistStudentDetailsPage, {
+        student: student,
+        status: this.studentList.get(student).location,
+        id: this.id
+      });
+    }
   }
 
   isEmpty() {
@@ -77,21 +79,23 @@ export class TherapistPage {
   //remove a student from therapist faorite ids
   //takes as input the student id for the student to be removed from the list
   removeStudent(SID: String){
-    //ask the user provider to remove the student from the favorited list
-    this.therapistStudents.splice(this.therapistStudents.indexOf(SID.toString()), 1);
+    if(this.editDesktop) {
+      //ask the user provider to remove the student from the favorited list
+      this.therapistStudents.splice(this.therapistStudents.indexOf(SID.toString()), 1);
 
-    this.userService.removeTherapistFavoriteID(String(this.id), SID).then((success) => {
-      if(success){
-        //now that the student is removed, update the list
-        this.userService.getTherapistFavoriteIDs(this.id.toString()).then((result:any) => {
-          this.therapistStudents = result;
-        }).catch((err) => {
-          console.log(err);
-        });
-      }
-    }).catch(err => {
-      console.log(err);
-    });
+      this.userService.removeTherapistFavoriteID(String(this.id), SID).then((success) => {
+        if(success){
+          //now that the student is removed, update the list
+          this.userService.getTherapistFavoriteIDs(this.id.toString()).then((result:any) => {
+            this.therapistStudents = result;
+          }).catch((err) => {
+            console.log(err);
+          });
+        }
+      }).catch(err => {
+        console.log(err);
+      });
+    }
   }
 
   edit() {
